+++ integration/serve/local.env
@@ -10,4 +10,3 @@ modal.state.modelWeightsOptions="pretrained"
 #"pretrained" "custom"
 modal.state.selectedModel="YOLOv5s"
 modal.state.weightsPath=""
-modal.state.device="cpu"
