+++ integration/serve/src/nn_utils.py
@@ -1,0 +1,38 @@
+import torch
+import numpy as np
+import supervisely as sly
+from supervisely.io.fs import get_file_name_with_ext
+import os
+from pathlib import Path
+import yaml
+import torch.nn.functional as F
+import cv2
+import itertools
+
+from utils.torch_utils import select_device
+from models.experimental import attempt_load
+from utils.general import check_img_size, non_max_suppression, scale_coords, xywh2xyxy
+from utils.datasets import letterbox
+from supervisely.geometry.sliding_windows_fuzzy import SlidingWindowsFuzzy, SlidingWindowBorderStrategy
+
+
+def construct_model_meta(model):
+    names = model.module.names if hasattr(model, 'module') else model.names
+
+    colors = None
+    if hasattr(model, 'module') and hasattr(model.module, 'colors'):
+        colors = model.module.colors
+    elif hasattr(model, 'colors'):
+        colors = model.colors
+    else:
+        colors = []
+        for i in range(len(names)):
+            colors.append(sly.color.generate_rgb(exist_colors=colors))
+
+    obj_classes = [sly.ObjClass(name, sly.Rectangle, color) for name, color in zip(names, colors)]
+    tags = [sly.TagMeta(CONFIDENCE, sly.TagValueType.ANY_NUMBER)]
+
+    return sly.ProjectMeta(obj_classes=sly.ObjClassCollection(obj_classes),
+                           tag_metas=sly.TagMetaCollection(tags))
+
+
