+++ integration/train/config.json
@@ -12,8 +12,8 @@
   "description": "Dashboard to configure and monitor training",
   "docker_image": "supervisely/yolov5:1.0.0",
   "instance_version": "6.4.57",
-  "main_script": "supervisely/train/src/sly_train.py",
-  "gui_template": "supervisely/train/src/gui.html",
+  "main_script": "integration/train/src/sly_train.py",
+  "gui_template": "integration/train/src/gui.html",
   "task_location": "workspace_tasks",
   "need_gpu": true,
   "icon": "https://i.imgur.com/HZCM0QM.png",
