+++ integration/train/src/sly_train_globals.py
@@ -18,8 +18,8 @@ ui_sources_dir = os.path.join(source_path, "ui")
 sys.path.append(ui_sources_dir)
 sly.logger.info(f"Added to sys.path: {ui_sources_dir}")
 
-load_dotenv(os.path.join(root_source_dir, "supervisely", "train", "debug.env"))
-load_dotenv(os.path.join(root_source_dir, "supervisely", "train", "secret_debug.env"), override=True)
+load_dotenv(os.path.join(root_source_dir, "integration", "train", "debug.env"))
+load_dotenv(os.path.join(root_source_dir, "integration", "train", "secret_debug.env"), override=True)
 
 my_app = AppService()
 my_app._ignore_stop_for_debug = True
