+++ integration/train/src/ui/architectures.html
