+++ integration/train/src/ui/artifacts.html
