+++ integration/train/src/ui/input_project.html
