+++ integration/train/src/ui/splits.html
