--- supervisely/train/secret_debug_example.env
@@ -1,3 +1,0 @@
-SERVER_ADDRESS="xxx"
-API_TOKEN="xxx"
-AGENT_TOKEN="xxx"
